factor out code related to the grammar into a separate module
